refactor: change dx and dy types from uint32_t to int8_t in InputData for better precision

+++ client/src/networks/protocol/factory/PacketFactory.hpp
@@ -82,7 +82,7 @@ namespace Network
          * @param shooting Shooting action flag
          * @return A shared pointer to the created packet
          */
-        std::shared_ptr<Net::IPacket> makeInput(uint32_t entity, float dx, float dy, uint8_t shooting) const noexcept;
+        std::shared_ptr<Net::IPacket> makeInput(uint32_t entity, int8_t dx, int8_t dy, uint8_t shooting) const noexcept;
 
       private:
         /**
